Refactor and update of Alist component to CityList, utils helper function moved to component CityList, db updated to improve readability, add more tests to Utis
