+++ pages/chapter4.tsx
@@ -2,7 +2,7 @@ import Layout from '@/components/Layout';
 import PageCard from '@/components/PageCard';
 import JSumRange from '@/components/exercises/chapter-4-jon/JSumRange';
 import ReversingAnArray from '@/components/exercises/chapter-4-oscar/ReversingAnArray';
-import Alist from '@/components/exercises/chapter-4-oscar/Alist';
+import CityList from '@/components/exercises/chapter-4-oscar/CityList';
 import SumOfARange from '@/components/exercises/chapter-4-oscar/SumOfARange';
 import chapter from '@/constants/chapter.json';
 import '../app/globals.css';
@@ -49,7 +49,7 @@ export default function Chapter4() {
                   id={devOscar[2].id}
                   title={devOscar[2].title}
                   details={devOscar[2].details}
-                  code={() => <Alist />}
+                  code={() => <CityList />}
                 />
               </>
             )}
